requirements.txt
requests>=2.25.1
schedule>=1.1.0
python-dotenv>=0.19.1

